% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Ntcrossprod}
\alias{example_Ntcrossprod}
\title{Example data for Ntcrossprod function}
\format{A numeric matrix with 100 rows and 10,000 cols:}
\usage{
example_Ntcrossprod
}
\description{
A dataset containing a numeric matrix with 1 and 0 as values
}
\seealso{
\code{\link{Ntcrossprod}}
}
\keyword{datasets}
