% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ntcrossprod.R
\name{Ntcrossprod}
\alias{Ntcrossprod}
\title{\code{tcrossprod} for narrow matrix.}
\usage{
Ntcrossprod(x, split_interval = 1000)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
The 'tcrossprod' result.
}
\description{
This function turns the \code{base::tcrossprod} less RAM memory intense for a narrow matrix
, i.e. the number of rows is much smaller than the number of columns. The function splits
the multiplication in several smaller \code{tcrossprod}s and summing them up at the end.
}
\examples{
#-Example-1-#
x <- matrix(sample(c(1,0),100000,replace=TRUE),100,10000)
Ntcrossprod(x)

#-Example-2-#
data(example_Ntcrossprod)
Ntcrossprod(example_Ntcrossprod)

}
