% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Ntcrossprod.R
\name{plot.Ntcrossprod}
\alias{plot.Ntcrossprod}
\title{\code{plot} Ntcrossprod class}
\usage{
\method{plot}{Ntcrossprod}(x, interval = 1:10, ...)
}
\arguments{
\item{x}{A Ntcrossprod class}

\item{interval}{sequence of the lines wish to plot}

\item{...}{parameters to be passed on corrplot function}
}
\value{
The corrplot graphic
}
\description{
Plots a corrplot for Ntcrossprod class product
}
\examples{
#-Example-1-#
data("example_Ntcrossprod")
y<-Ntcrossprod(example_Ntcrossprod)
plot(y)

#-Example-2-#
data("example_Ntcrossprod")
y<-Ntcrossprod(example_Ntcrossprod)
plot(y, interval=1:5, method="square")

}
